pyproject.toml
[tool.poetry]
name = "agent1"
version = "0.1.0"
description = "LangChain Agent with ReAct"
authors = ["Vishnu Sangadala"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
langchain = "==0.3.*"
langchain-openai = "==0.2.*"
python-dotenv = "^1.0.0"
langchain-core = "==0.3.*"
langchain-community = "==0.3.*"
langchainhub = "^0.1.21"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"